#28 move auction slider components to another folder, settle outstanding auction view
