+++ src/app/auction-slider/outstanding-auction.component.css
@@ -1,0 +1,3 @@
+.outstanding--image {
+    text-align: center;
+}
