--- src/app/auction/auction-slider.component.html
@@ -1,3 +1,0 @@
-<div *ngIf="auctions.length > 0" class="row">
-  <auction *ngFor="let auction of auctions" [auction]="auction"></auction>
-</div>
